refactor(jsonb-query): group dialect modules under src/dialect/
The package had grown to 11 flat source modules. Move the dialect
cluster — base contract, legacy + jsonpath dialects, and the
jsonpath-only escape helpers — into src/dialect/ with its own barrel,
leaving 7 focused modules at the root. Internal-only: index.ts exports
and package.json `exports` are unchanged, so the published API is
identical (124 tests pass, build unchanged). No changeset needed.
