+++ packages/jsonb-query/src/dialect/index.ts
@@ -1,0 +1,3 @@
+export * from './base';
+export { legacyDialect } from './legacy';
+export { jsonpathDialect } from './jsonpath';
